+++ Loja/Loja.Application/Services/ClienteService.cs
@@ -1,6 +1,4 @@
-﻿using AutoMapper;
-using Loja.Application.DTOs;
-using Loja.Application.Interfaces;
+﻿using Loja.Application.Interfaces;
 using Loja.Domain.Entities;
 using Loja.Domain.Interfaces;
 using Loja.Domain.PaginationEntities;
@@ -10,24 +8,21 @@ namespace Loja.Application.Services
     public class ClienteService : IClienteService
     {
         private readonly IClienteRepository _clienteRepository;
-        private readonly IMapper _mapper;
-        public ClienteService(IMapper mapper, IClienteRepository clienteRepository)
+        public ClienteService(IClienteRepository clienteRepository)
         {
             _clienteRepository = clienteRepository ??
                 throw new ArgumentNullException(nameof(clienteRepository));
-
-            _mapper = mapper;
         }
 
         /// <summary>
         /// Obtém um cliente pelo Id.
         /// </summary>
         /// <param name="id">Id do cliente.</param>
-        /// <returns>Retorna um ClienteDTO</returns>
-        public async Task<ClienteDTO> GetById(int? id)
+        /// <returns>Retorna um Cliente</returns>
+        public async Task<Cliente> GetById(int? id)
         {
             var clienteEntity = await _clienteRepository.GetByIdAsync(x => x.Id == id);
-            return _mapper.Map<ClienteDTO>(clienteEntity);
+            return clienteEntity;
         }
 
         /// <summary>
@@ -35,49 +30,45 @@ namespace Loja.Application.Services
         /// </summary>
         /// <param name="parameters">Objeto com os parâmetros de paginação</param>
         /// <returns>Retorna um objeto PagingList com a lista de clientes e os dados de paginação</returns>
-        public async Task<PagingList<ClienteDTO>> GetClientes(PagingParameters parameters)
+        public async Task<PagingList<Cliente>> GetClientes(PagingParameters parameters)
         {
             var pagingList = await _clienteRepository.GetAsync(parameters);
 
-            var clientesDto = _mapper.Map<List<ClienteDTO>>(pagingList.Items);
-
-            var pagingListDto = new PagingList<ClienteDTO>(clientesDto, pagingList.PaginationInfo);
-
-            return pagingListDto;
+            return pagingList;
         }
 
         /// <summary>
         /// Adiciona um cliente à tabela do banco de dados.
         /// </summary>
-        /// <param name="clienteDto">Objeto com os dados do cliente a ser adicionado.</param>
+        /// <param name="clienteNovo">Objeto com os dados do cliente a ser adicionado.</param>
         /// <returns>Retorna o cliente adicionado.</returns>
-        public async Task<ClienteDTO> Add(ClienteDTO clienteDto)
+        public async Task<Cliente> Add(Cliente clienteNovo)
         {
-            var clienteEntity = _mapper.Map<Cliente>(clienteDto);
-            var cliente = await _clienteRepository.CreateAsync(clienteEntity);
-            return _mapper.Map<ClienteDTO>(cliente);
+            var cliente = await _clienteRepository.CreateAsync(clienteNovo);
+            return cliente;
         }
 
         /// <summary>
         /// Remove um cliente do banco de dados.
         /// </summary>
         /// <param name="id">Id do cliente.</param>
-        /// <returns></returns>
         public async Task Remove(int? id)
         {
             var clienteEntity = _clienteRepository.GetByIdAsync(x => x.Id == id).Result;
+
+            if (clienteEntity == null)
+                return;
+
             await _clienteRepository.RemoveAsync(clienteEntity);
         }
 
         /// <summary>
         /// Atualiza um cliente.
         /// </summary>
-        /// <param name="clienteDto">Objeto com os dados do cliente a ser atualizado.</param>
-        /// <returns></returns>
-        public async Task Update(ClienteDTO clienteDto)
+        /// <param name="cliente">Objeto com os dados do cliente a ser atualizado.</param>
+        public async Task Update(Cliente cliente)
         {
-            var clienteEntity = _mapper.Map<Cliente>(clienteDto);
-            await _clienteRepository.UpdateAsync(clienteEntity);
+            await _clienteRepository.UpdateAsync(cliente);
         }
     }
 }
