+++ Loja/Loja.Application/Services/ProdutoService.cs
@@ -1,6 +1,4 @@
-﻿using AutoMapper;
-using Loja.Application.DTOs;
-using Loja.Application.Interfaces;
+﻿using Loja.Application.Interfaces;
 using Loja.Domain.Entities;
 using Loja.Domain.Interfaces;
 using Loja.Domain.PaginationEntities;
@@ -11,13 +9,10 @@ namespace Loja.Application.Services
     public class ProdutoService : IProdutoService
     {
         private readonly IProdutoRepository _produtoRepository;
-        private readonly IMapper _mapper;
-        public ProdutoService(IMapper mapper, IProdutoRepository produtoRepository)
+        public ProdutoService(IProdutoRepository produtoRepository)
         {
             _produtoRepository = produtoRepository ??
                  throw new ArgumentNullException(nameof(produtoRepository));
-
-            _mapper = mapper;
         }
 
         /// <summary>
@@ -25,15 +20,11 @@ namespace Loja.Application.Services
         /// </summary>
         /// <param name="parameters">Objeto com os parâmetros de paginação</param>
         /// <returns>Retorna um objeto PagingList com a lista de produtos e os dados de paginação</returns>
-        public async Task<PagingList<ProdutoDTO>> GetProdutos(PagingParameters parameters)
+        public async Task<PagingList<Produto>> GetProdutos(PagingParameters parameters)
         {
             var pagingList = await _produtoRepository.GetAsync(parameters);
 
-            var produtosDto = _mapper.Map<List<ProdutoDTO>>(pagingList.Items);
-
-            var pagingListDto = new PagingList<ProdutoDTO>(produtosDto, pagingList.PaginationInfo);
-
-            return pagingListDto;
+            return pagingList;
         }
 
 
@@ -43,100 +34,65 @@ namespace Loja.Application.Services
         /// <param name="parameters">Objeto com os parâmetros de paginação</param>
         /// <param name="predicate">Delegate com o critério de busca.</param>
         /// <returns>Retorna um objeto PagingList com a lista de produtos e os dados de paginação</returns>
-        public async Task<PagingList<ProdutoDTO>> GetProdutos(PagingParameters parameters, Expression<Func<Produto, bool>> predicate)
+        public async Task<PagingList<Produto>> GetProdutos(PagingParameters parameters, Expression<Func<Produto, bool>> predicate)
         {
             var pagingList = await _produtoRepository.GetAsync(parameters, predicate);
 
-            var produtosDto = _mapper.Map<List<ProdutoDTO>>(pagingList.Items);
-
-            var pagingListDto = new PagingList<ProdutoDTO>(produtosDto, pagingList.PaginationInfo);
-
-            return pagingListDto;
+            return pagingList;
         }
 
         /// <summary>
         /// Obtém um produto pelo Id.
         /// </summary>
         /// <param name="id">Id do produto</param>
-        /// <returns>Retorna um objeto ProdutoDTO</returns>
-        public async Task<ProdutoDTO> GetById(int? id)
+        /// <returns>Retorna um objeto Produto</returns>
+        public async Task<Produto> GetById(int? id)
         {
             var produtoEntity = await _produtoRepository.GetByIdAsync(x => x.Id == id);
-            return _mapper.Map<ProdutoDTO>(produtoEntity);
+            return produtoEntity;
         }
 
         /// <summary>
         /// Obtém um produto pelo Id com a sua categoria.
         /// </summary>
         /// <param name="id">Id do produto</param>
-        /// <returns>Retorna um objeto ProdutoDTO com a Categoria inclusa.</returns>
-        public async Task<ProdutoDTO> GetByIdIncludeCategoria(int id)
+        /// <returns>Retorna um objeto Produto com a Categoria inclusa.</returns>
+        public async Task<Produto> GetByIdIncludeCategoria(int id)
         {
             var produtoEntity = await _produtoRepository.GetByIdIncludeCategoriaAsync(id);
-            return _mapper.Map<ProdutoDTO>(produtoEntity);
+            return produtoEntity;
         }
 
         /// <summary>
         /// Adiciona um produto à tabela do banco de dados.
         /// </summary>
-        /// <param name="produtoDto">Objeto com os dados do produto a ser adicionado.</param>
+        /// <param name="produtoNovo">Objeto com os dados do produto a ser adicionado.</param>
         /// <returns>Retorna o produto adicionado.</returns>
-        public async Task<ProdutoDTO> Add(ProdutoDTO produtoDto)
+        public async Task<Produto> Add(Produto produtoNovo)
         {
-            var produtoEntity = _mapper.Map<Produto>(produtoDto);
-            var produto = await _produtoRepository.CreateAsync(produtoEntity);
-            return _mapper.Map<ProdutoDTO>(produto);
+            var produto = await _produtoRepository.CreateAsync(produtoNovo);
+            return produto;
         }
 
         /// <summary>
         /// Atualiza um produto.
         /// </summary>
-        /// <param name="produtoDto">Objeto com os dados do produto a ser atualizado.</param>
-        /// <returns></returns>
-        public async Task Update(ProdutoDTO produtoDto)
+        /// <param name="produto">Objeto com os dados do produto a ser atualizado.</param>
+        public async Task Update(Produto produto)
         {
-            var produtoEntity = _mapper.Map<Produto>(produtoDto);
-            await _produtoRepository.UpdateAsync(produtoEntity);
+            await _produtoRepository.UpdateAsync(produto);
         }
 
         /// <summary>
         /// Remove um produto do banco de dados.
         /// </summary>
         /// <param name="id">Id do produto.</param>
-        /// <returns></returns>
         public async Task Remove(int? id)
         {
             var produtoEntity = _produtoRepository.GetByIdAsync(x => x.Id == id).Result;
+            if (produtoEntity == null)
+                return;
             await _produtoRepository.RemoveAsync(produtoEntity);
-        }
-
-        /// <summary>
-        /// Converte a string da ordenação para uma expressão lambda.
-        /// </summary>
-        /// <param name="orderedBy">String com a palavra da ordenação.</param>
-        /// <returns>Retorna a expressão lambda a ser usada para ordenar.</returns>
-        private static Expression<Func<Produto, object>> SwitchCaseOrderedBy(string orderedBy)
-        {
-            switch (orderedBy.ToLower())
-            {
-                case "name":
-                case "nome":
-                    return x => x.Nome;
-                case "preco":
-                case "price":
-                case "preço":
-                    return x => x.Preco;
-                case "catid":
-                case "categoriaid":
-                    return x => x.CategoriaId;
-                case "estoque":
-                    return x => x.Estoque;
-                case "data":
-                case "date":
-                    return x => x.DataCadastro;
-                default:
-                    return x => x.Id;
-            }
         }
     }
 }
